% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeUnsupervisedMetrics.R
\name{computeUnsupervisedMetrics}
\alias{computeUnsupervisedMetrics}
\title{computeUnsupervisedMetrics}
\usage{
computeUnsupervisedMetrics(sce, clusters)
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{clusters}{Clustering results for the SingleCellExperiment object}
}
\value{
unsupervised metrics
}
\description{
computeUnsupervisedMetrics
}
