% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{embryoClusts}
\alias{embryoClusts}
\title{Embryo data clustering results from two pipelines.}
\format{
A list with two levels
}
\usage{
embryoClusts
}
\description{
Clustering results of the embryo scRNA-seq data from two pipelines. Pipelines computed are the default pipeline settings from runPipelineCombs.
Structured as a list with two entries, each entry is a factor that indicates the cluster that each of the cells in the embryo datasets belong to.
}
\keyword{datasets}
