% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSetEval.R
\name{geneSetEval}
\alias{geneSetEval}
\title{geneSetEval}
\usage{
geneSetEval(sce, clusters)
}
\arguments{
\item{sce}{A SingleCellExperiment object with raw counts in the counts slot.}

\item{clusters}{Factor representing clustering results from a pipeline for the SingleCellExperiment.}
}
\value{
Tibble containing gene set enrichment results for all clusters.
}
\description{
A function that uses MSigDB Hallmark Pathways to compute gene set enrichment analysis for each clustering output
given in clusters.
}
\examples{
# Example:
\dontrun{
data(embryo)
data(embryoClusts)
# Compute pathways for the first clustering output
geneSetRes <- geneSetEval(embryo, embryoClusts[[1]])
geneSetRes
}
}
