% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gsea}
\alias{gsea}
\title{GSEA results for embryo data clustered with first pipeline option.}
\format{
A list with 4 levels.
}
\usage{
gsea
}
\description{
GSEA results for embryo data clustered with first pipeline option.
}
\keyword{datasets}
