% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{h.all.v7.4.symbols.gmt}
\alias{h.all.v7.4.symbols.gmt}
\title{MSigDB Hallmark Pathways for gene set enrichment analysis}
\format{
pathways
}
\usage{
h.all.v7.4.symbols.gmt
}
\description{
MSigDB Hallmark Pathways for gene set enrichment analysis
}
\keyword{datasets}
