% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runClustREval.R
\name{runClustREval}
\alias{runClustREval}
\title{Launch Shiny App for clustREval}
\usage{
runClustREval()
}
\value{

}
\description{
A function that launches the Shiny app for clustREval
The code has been placed in \code{./inst/shiny-scripts}.
}
