% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPipelineCombs.R
\name{runPipelineCombs}
\alias{runPipelineCombs}
\title{runPipelineCombs}
\usage{
runPipelineCombs(
  sce,
  outputPrefix = "sce",
  doubletmethod = c("none"),
  filt = c("filt.default"),
  norm = c("norm.seurat"),
  sel = c("sel.vst"),
  selnb = 2000,
  dr = c("seurat.pca"),
  clustmethod = c("clust.seurat"),
  dims = c(10),
  resolution = c(0.1, 0.2)
)
}
\arguments{
\item{sce}{SingleCellExperiment object with raw counts in counts slot.}

\item{outputPrefix}{String indicating the prefix of the filepath that results should save to}

\item{doubletmethod}{Character list of doublet detection methods}

\item{filt}{Character list of filtering methods}

\item{norm}{Character list of normalization methods}

\item{sel}{Character list of selection methods}

\item{selnb}{Character list of number of genes to select}

\item{dr}{Character list of dimensionality reduction methods}

\item{clustmethod}{Character list of clustering methods}

\item{dims}{Numeric list of dimensions for dimensionality reduction}

\item{resolution}{Character list of clustering resolutions}
}
\value{
List of factors containing clustering results from each pipeline run.
}
\description{
A function that runs user-defined clustering pipelines by taking all possible combinations of user-provided pipeline hyperparameters.
User-provided hyperparameters must match available options.
}
