% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeUnsupervisedMetrics.R
\name{computeUnsupervisedMetrics}
\alias{computeUnsupervisedMetrics}
\title{computeUnsupervisedMetrics}
\usage{
computeUnsupervisedMetrics(sce, clusters)
}
\arguments{
\item{sce}{A SingleCellExperiment object with raw counts in the counts slot.}

\item{clusters}{Factor representing clustering results for the SingleCellExperiment object.}
}
\value{
A list with the Dunn Index and mean Silhouette score as elements
}
\description{
A function that computes Dunn Index and and mean Silhouette score for a clustering output
}
\examples{
# Example:

}
