% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{embryo}
\alias{embryo}
\title{Tracing pluripotency of human early embryos and embryonic stem cells by single cell RNA-seq}
\format{
A data frame with 26073 rows and 124 columns:
}
\source{
\url https://www.ebi.ac.uk/gxa/sc/experiments/E-GEOD-36552/results/tsne
}
\usage{
embryo
}
\description{
Gene expression of 124 early embryo and embryonic stem cells from scRNA-seq. Sourced from the
EMBL-EBI Single Cell Expression Atlas.
}
\keyword{datasets}
