% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSetEval.R
\name{geneSetEval}
\alias{geneSetEval}
\title{geneSetEval}
\usage{
geneSetEval(sce, clusters)
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{clusters}{Clustering results from the SingleCellExperiment}
}
\value{
Mean enrichment score for genes in each cluster
}
\description{
geneSetEval
}
