% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneSetEval.R
\name{plotGeneSetEval}
\alias{plotGeneSetEval}
\title{plotGeneSetEval}
\usage{
plotGeneSetEval(sce, clusters, gseas)
}
\arguments{
\item{sce}{A single cell experiment object}

\item{clusters}{Clustering results for the SingleCellExperiment}

\item{gseas}{A tibble containing enrichment values for each cluster computed from geneSetEval}
}
\value{
a plot
}
\description{
A function that plots gene set enrichment analysis results for each cluster in
order to visualize biological differences between clusters.
}
