% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneSetEval.R
\name{plotGeneSetEval}
\alias{plotGeneSetEval}
\title{Plot Enrichment Scores for Each Pathway in a Cluster}
\usage{
plotGeneSetEval(gseas, clustname = "")
}
\arguments{
\item{gseas}{A list containing enrichment values for each pathway in a cluster}

\item{clustname}{A string or integer containing the name of the cluster. Used in the title of the plot.}
}
\value{
A plot for one cluster from a clustering output showing the enriched pathways for that cluster.
}
\description{
A function that plots gene set enrichment analysis results for a cluster in
order to visualize biological differences between clusters.
}
\examples{
# Example:
\dontrun{
data(embryo)
data(embryoClusts)
# Compute pathways for the first clustering output
geneSetRes <- geneSetEval(embryo, embryoClusts[[1]])
plotGeneSetEval(geneSetRes[[1]])
}

}
