% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneSetEval.R
\name{plotGeneSetEval}
\alias{plotGeneSetEval}
\title{plotGeneSetEval}
\usage{
plotGeneSetEval(gseas)
}
\arguments{
\item{gseas}{A list containing enrichment values for each cluster computed from geneSetEval}
}
\value{
A plot for each clustering output showing the enriched pathways for each.
}
\description{
A function that plots gene set enrichment analysis results for each cluster in
order to visualize biological differences between clusters.
}
\examples{
# Example:
\dontrun{
data(embryo)
data(embryoClusts)
# Compute pathways for the first clustering output
geneSetRes <- geneSetEval(embryo, embryoClusts[[1]])
plotGeneSetEval(geneSetRes)
}
}
