% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPipelineCombs.R
\name{runPipelineCombs}
\alias{runPipelineCombs}
\title{runPipelineCombs}
\usage{
runPipelineCombs(sce, filt, norm, resolution)
}
\arguments{
\item{sce}{SingleCellExperiment object with counts matrix and phenoid colData entry.}

\item{filt}{list of filtering methods}

\item{norm}{list of normalization methods}

\item{resolution}{list of clustering resolutions}
}
\value{
end outputs object
}
\description{
Run user-defined clustering pipeline combinations on a user-specificed scRNA-seq dataset.
}
