% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPipelineCombs.R
\name{runPipelineCombs}
\alias{runPipelineCombs}
\title{runPipelineCombs}
\usage{
runPipelineCombs(
  sce,
  outputPrefix = "sce",
  doubletmethod = c("none"),
  filt = c("filt.default"),
  norm = c("norm.seurat"),
  sel = c("sel.vst"),
  selnb = 2000,
  dr = c("seurat.pca"),
  clustmethod = c("clust.seurat"),
  dims = c(10),
  resolution = c(0.1, 0.2)
)
}
\arguments{
\item{sce}{SingleCellExperiment object with raw counts in counts slot.}

\item{outputPrefix}{String indicating the prefix of the filepath that results should save to}

\item{doubletmethod}{Character vector of doublet detection methods. Default: none}

\item{filt}{Character vector of filtering methods. Default: filt.default}

\item{norm}{Character vector of normalization methods Default: norm.seurat}

\item{sel}{Character vector of selection methods. Default: sel.vst}

\item{selnb}{Numeric vector of number of genes to select. Default: 2000}

\item{dr}{Character vector of dimensionality reduction methods. Default: seurat.pca}

\item{clustmethod}{Character vector of clustering methods Default: clust.seurat}

\item{dims}{Numeric vector of dimensions for dimensionality reduction. Default: 10}

\item{resolution}{Character vector of clustering resolutions. Default: 0.1, 0.2}
}
\value{
List of factors containing clustering results from each pipeline run.
}
\description{
A function that runs user-defined clustering pipelines by taking all possible combinations of user-provided pipeline hyperparameters.
User-provided hyperparameters must match available options.
}
\details{
Valid filtering methods:
filt.default, filt.lenient, filt.mad, filt.pca, filt.pca2, filt.stringent, none

Valid normalization methods:
norm.none, norm.none.scaled, norm.scnorm, norm.scnorm.scaled, norm.scran, norm.scran.scaled,
norm.sctransform, norm.scVI, norm.seurat, norm.seuratvst

Valid feature selection methods:
sel.deviance, sel.expr, sel.fromField, sel.vst

Valid dimensionality reduction method:
seurat.pca, serat.pca.noweight
}
\examples{
# Example 1: Run with default parameters
data(embryo)
res <- clustREval::runPipelineCombs(sce=embryo, outputPrefix="embryo")
res

# Example 2: Run with user-defined parameters
data(embryo)
filtMethods <- c("filt.lenient", "filt.stringent")
res <- runPipelineCombs(sce=embryo, outputPrefix="embryo", filt=filtMethods)
res
}
